Refactor: createHabitData 로직 분리
